Revert "Merge pull request #45 from UNIZAR-30226-2026-07/feat/arreglos-fixes-movimiento-vision"
This reverts commit da96e4c82830c64bdf9f12739bcaa159b24998c4, reversing
changes made to 61f10777db54406b11dacb8b0aa6f8da8e92cac7.
