+++ .gitignore
@@ -137,5 +137,3 @@ dist
 # Vite logs files
 vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
-
-docs/asyncapi/output
