+++ src/config/gameRules.js
@@ -1,10 +1,6 @@
 /**
  * Configuración Global de Reglas de Juego
- * Centraliza los valores de balanceo para facilitar ajustes sin tocar la lógica del motor.
- * @constant {Object} GAME_RULES
- * @property {Object} MAP - Configuración del tablero.
- * @property {Object} RESOURCES - Gestión de costes y regeneración de combustible/munición.
- * @property {Object} COMBAT - Valores de daño y persistencia de proyectiles.
+ * Centraliza los valores de balanceo para facilitar ajustes sin tocar la lógica.
  */
 export const GAME_RULES = {
     MAP: {
@@ -12,9 +8,9 @@ export const GAME_RULES = {
         DEPLOY_ZONE_Y: 4
     },
     RESOURCES: {
-        MAX_FUEL: 10,
-        RESET_FUEL: 10,
-        RESET_AMMO: 10,
+        MAX_FUEL: 30,
+        REGEN_FUEL: 10,
+        RESET_AMMO: 5,
         COST_MOVE: 1,
         COST_ROTATE: 2,
         COST_CANNON: 2,
