+++ src/modules/engine/services/engineService.js
@@ -1,11 +1,13 @@
 /**
  * Servicio de Lógica del Motor (EngineService)
- * Gestiona movimientos, rotaciones, colisiones y límites.
+ * Gestiona movimientos, rotaciones y validaciones de límites de tablero.
  */
 import { GAME_RULES } from '../../../config/gameRules.js';
 
 /**
  * Calcula la traslación de coordenadas
+ * @param {Object} pos - {x, y}
+ * @param {string} dir - N, S, E, W
  */
 export const calcularTraslacion = (pos, dir) => {
     let { x, y } = pos;
@@ -17,7 +19,7 @@ export const calcularTraslacion = (pos, dir) => {
 };
 
 /**
- * Valida si una posición está dentro del mapa 15x15
+ * Valida si una posición está dentro del mapa
  */
 export const validarLimitesMapa = (x, y) => {
     const limit = GAME_RULES.MAP.SIZE - 1;
@@ -25,7 +27,9 @@ export const validarLimitesMapa = (x, y) => {
 };
 
 /**
- * Calcula la nueva orientación cardinal
+ * Calcula la nueva orientación
+ * @param {string} actual - N, S, E, W
+ * @param {number} grados - 90, -90
  */
 export const calcularRotacion = (actual, grados) => {
     const orden = ['N', 'E', 'S', 'W'];
@@ -36,58 +40,11 @@ export const calcularRotacion = (actual, grados) => {
 };
 
 /**
- * Devuelve los costes estandarizados de la V1
+ * Devuelve los costes estandarizados
  */
 export const obtenerCostesMovimiento = () => {
     return {
         TRASLACION: GAME_RULES.RESOURCES.COST_MOVE,
         ROTACION: GAME_RULES.RESOURCES.COST_ROTATE
     };
-};
-
-/**
- * Calcula todas las celdas ocupadas por un barco basándose en su origen, orientación y tamaño
- */
-export const calcularCeldasOcupadas = (startX, startY, orientation, size) => {
-    const celdas = [];
-    const esHorizontal = orientation === 'E' || orientation === 'W';
-    
-    for (let i = 0; i < size; i++) {
-        celdas.push({
-            x: esHorizontal ? startX + i : startX,
-            y: esHorizontal ? startY : startY + i
-        });
-    }
-    return celdas;
-};
-
-/**
- * Verifica si las celdas objetivo intersectan con algún barco vivo
- */
-export const verificarColision = (targetCells, allAliveShips, ignoreShipId) => {
-    for (const ship of allAliveShips) {
-        if (ship.id === ignoreShipId) continue;
-        
-        const size = Math.max(ship.UserShip.ShipTemplate.width, ship.UserShip.ShipTemplate.height);
-        const occupiedCells = calcularCeldasOcupadas(ship.x, ship.y, ship.orientation, size);
-        
-        for (const tCell of targetCells) {
-            for (const oCell of occupiedCells) {
-                if (tCell.x === oCell.x && tCell.y === oCell.y) {
-                    return true; 
-                }
-            }
-        }
-    }
-    return false;
-};
-
-/**
- * Traduce comandos de dirección del jugador Sur (Vista invertida)
- */
-export const traducirDireccionEntrada = (inputDirection, playerSide) => {
-    if (playerSide === 'NORTH') return inputDirection;
-    
-    const invertY = { 'N': 'S', 'S': 'N', 'E': 'E', 'W': 'W' };
-    return invertY[inputDirection] || inputDirection;
 };
