+++ src/modules/engine/sockets/vision.int.test.js
@@ -1,26 +1,29 @@
 /**
  * Test de Integración: Niebla de Guerra y Emisión Asimétrica.
+ * Garantiza que red manda perspectivas diferentes a cada jugador, 
+ * preparándolo para la V2 (filtrado estricto).
  */
 import { createServer } from 'http';
 import { Server } from 'socket.io';
 import { io as Client } from 'socket.io-client';
-import { sequelize } from '../../../config/index.js';
-import { createCompleteMatch, socketProtect } from '../../../shared/index.js';
-import { authService } from '../../auth/index.js';
-import { registerGameHandlers } from '../../game/index.js';
-import { registerEngineHandlers } from '../index.js';
+import { sequelize } from '../../../config/index.js';          
+import { createCompleteMatch, socketProtect } from '../../../shared/index.js'; 
+import { authService } from '../../auth/index.js';              
+import { registerGameHandlers } from '../../game/index.js';  
+import { registerEngineHandlers } from '../index.js';  
 
 describe('Vision & Fog of War: Asymmetric Socket Responsibility', () => {
     let io, server, hostClient, guestClient, setup;
-    const port = 4238; 
+    const port = 4230; //Si cambiais el puerto y el test se rompe, cambiad el test
 
     beforeAll(async () => {
         await sequelize.query('DROP SCHEMA public CASCADE; CREATE SCHEMA public;');
         await sequelize.sync({ force: true });
 
+        // Factoría: Host(NORTH) barco en {5,5}. Guest(SOUTH) barco en {5,7}.
         setup = await createCompleteMatch(
-            { username: 'v_host', email: 'vh@test.com' },
-            { username: 'v_guest', email: 'vg@test.com' }
+            { username: 'vision_host', email: 'vh@t.va' },
+            { username: 'vision_guest', email: 'vg@t.va' }
         );
 
         server = createServer();
@@ -43,7 +46,16 @@ describe('Vision & Fog of War: Asymmetric Socket Responsibility', () => {
             new Promise(res => hostClient.on('connect', res)),
             new Promise(res => guestClient.on('connect', res))
         ]);
-    }, 20000);
+
+        // Unimos a ambos a la sala de la partida
+        hostClient.emit('game:join', setup.match.id);
+        guestClient.emit('game:join', setup.match.id);
+
+        await Promise.all([
+            new Promise(res => hostClient.once('game:joined', res)),
+            new Promise(res => guestClient.once('game:joined', res))
+        ]);
+    });
 
     afterAll(async () => {
         hostClient.disconnect();
@@ -53,57 +65,39 @@ describe('Vision & Fog of War: Asymmetric Socket Responsibility', () => {
         await sequelize.close();
     });
 
-    it('Debe emitir perspectivas independientes a cada jugador al moverse', (done) => {
-        let updatesReceived = 0;
-        let hostPayload = null;
-        let guestPayload = null;
-
-        // Configuramos la escucha antes de mover
-        hostClient.on('match:vision_update', (payload) => {
-            hostPayload = payload;
-            updatesReceived++;
-            if (updatesReceived === 2) validate();
+    it('Debe emitir perspectivas independientes a cada jugador al moverse (V1 / V2 Readiness)', async () => {
+        const hostPromise = new Promise(resolve => {
+            hostClient.once('match:vision_update', (payload) => resolve(payload));
         });
 
-        guestClient.on('match:vision_update', (payload) => {
-            guestPayload = payload;
-            updatesReceived++;
-            if (updatesReceived === 2) validate();
+        const guestPromise = new Promise(resolve => {
+            guestClient.once('match:vision_update', (payload) => resolve(payload));
         });
 
-        const validate = () => {
-            try {
-                // Host ve su barco movido a Y=6 absoluto
-                expect(hostPayload.myFleet[0].y).toBe(6);
-                // Guest ve su barco en Y=7 local (Y=7 absoluto)
-                expect(guestPayload.myFleet[0].y).toBe(7);
-                // Guest ve al enemigo en Y=8 local (Y=6 absoluto -> 14-6=8)
-                expect(guestPayload.visibleEnemyFleet[0].y).toBe(8);
-                done();
-            } catch (e) {
-                done(e);
-            }
-        };
+        // El host mueve su barco al Sur. Pasa de {5,5} a {5,6}.
+        hostClient.emit('ship:move', {
+            matchId: setup.match.id,
+            shipId: setup.shipH.id,
+            direction: 'S'
+        });
 
-        hostClient.emit('game:join', setup.match.id);
-        guestClient.emit('game:join', setup.match.id);
+        const [hostVision, guestVision] = await Promise.all([hostPromise, guestPromise]);
 
-        let joinedCount = 0;
-        const onJoined = () => {
-            joinedCount++;
-            if (joinedCount === 2) {
-                setTimeout(() => {
-                    hostClient.emit('ship:move', {
-                        matchId: setup.match.id,
-                        shipId: setup.shipH.id,
-                        direction: 'S'
-                    });
-                }, 100);
-            }
-        };
+        //EN   V2 ASEGURARSE QUE LO QUE VERIFICA ES CORRECTO (BARCOS SE VEN, ETC)
+        // VALIDACIÓN HOST (NORTH)
+        // El Host ve su barco donde lo movió ({5,6} mirando al 'N')
+        expect(hostVision.myFleet[0].y).toBe(6);
+        expect(hostVision.myFleet[0].orientation).toBe('N');
+        // El Host ve al enemigo en coordenadas absolutas ({5,7} mirando al 'S')
+        expect(hostVision.visibleEnemyFleet[0].y).toBe(7);
+        expect(hostVision.visibleEnemyFleet[0].orientation).toBe('S');
 
-        hostClient.once('game:joined', onJoined);
-        guestClient.once('game:joined', onJoined);
-        
-    }, 20000);
+        // VALIDACIÓN GUEST (SOUTH)
+        // El Guest ve su barco en perspectiva local: Absoluto 7 -> 14 - 7 = 7. Orientación Absoluta 'S' -> 'N'
+        expect(guestVision.myFleet[0].y).toBe(7);
+        expect(guestVision.myFleet[0].orientation).toBe('N');
+        // El Guest ve el barco enemigo en perspectiva local: Absoluto 6 -> 14 - 6 = 8. Orientación Absoluta 'N' -> 'S'
+        expect(guestVision.visibleEnemyFleet[0].y).toBe(8);
+        expect(guestVision.visibleEnemyFleet[0].orientation).toBe('S');
+    });
 });
