+++ src/modules/game/models/MatchPlayer.test.js
@@ -6,7 +6,7 @@ import MatchPlayer from './MatchPlayer.js';
 describe('MatchPlayer Model Unit Tests', () => {
     it('Debe tener recursos iniciales correctos según GAME_RULES V1', () => {
         expect(MatchPlayer.rawAttributes.fuelReserve.defaultValue).toBe(10);
-        expect(MatchPlayer.rawAttributes.ammoCurrent.defaultValue).toBe(10);
+        expect(MatchPlayer.rawAttributes.ammoCurrent.defaultValue).toBe(5);
     });
 
     it('No debe permitir combustible negativo', async () => {
