+++ src/modules/game/services/index.js
@@ -1,10 +1,10 @@
-import * as boardUtils from './boardUtils.js';
+/**
+ * Fachada de servicios del módulo Game.
+ */
 import * as matchService from './matchService.js';
 import * as statusService from './matchStatusService.js';
-import * as visionService from './visionService.js';
 
 export {
-    boardUtils, matchService,
-    statusService,
-    visionService
+    matchService,
+    statusService
 };
