+++ src/modules/game/services/matchService.test.js
@@ -1,7 +1,4 @@
-/**
- * Test Unitario: Lógica de Servicio de Partida
- */
-import { calcularRegeneracionTurno, traducirOrientacion, traducirPosicionTablero } from './matchService.js';
+import { calcularRegeneracionTurno, traducirPosicionTablero, traducirOrientacion } from './matchService.js';
 
 describe('MatchService Unit Tests (Pure Logic & Translation)', () => {
 
@@ -11,18 +8,23 @@ describe('MatchService Unit Tests (Pure Logic & Translation)', () => {
         const southPos = traducirPosicionTablero(relativePos, 'SOUTH');
 
         expect(northPos).toEqual({ x: 5, y: 0 });
-        expect(southPos).toEqual({ x: 5, y: 14 }); 
+        expect(southPos).toEqual({ x: 5, y: 14 }); // 15 - 1 - 0 = 14
     });
 
     it('traducirOrientacion - Debe invertir Norte/Sur para el bando SOUTH', () => {
         expect(traducirOrientacion('N', 'NORTH')).toBe('N');
         expect(traducirOrientacion('N', 'SOUTH')).toBe('S');
-        expect(traducirOrientacion('E', 'SOUTH')).toBe('E');
+        
+        expect(traducirOrientacion('E', 'SOUTH')).toBe('E'); // El Este y Oeste no se invierten
     });
 
-    it('calcularRegeneracionTurno - Debe rellenar AP y MP a 10 según nuevas reglas', () => {
-        const res1 = calcularRegeneracionTurno({ fuel: 2, ammo: 0 });
-        expect(res1.fuel).toBe(10);
-        expect(res1.ammo).toBe(10);
+    it('calcularRegeneracionTurno - Debe rellenar AP y sumar MP limitados a 30', () => {
+        const res1 = calcularRegeneracionTurno({ fuel: 10, ammo: 0 });
+        expect(res1.fuel).toBe(20);
+        expect(res1.ammo).toBe(5);
+
+        const res2 = calcularRegeneracionTurno({ fuel: 25, ammo: 3 });
+        expect(res2.fuel).toBe(30); // Cap at 30
+        expect(res2.ammo).toBe(5);
     });
 });
